+++ app/Main.hs
@@ -19,8 +19,10 @@ import           Data.Extensible
 import           Data.Extensible.GetOpt
 import qualified Data.Yaml              as Y
 import           GetOpt                 (withGetOpt')
+import qualified Git
 import           Mix
 import           Mix.Plugin.Logger      as MixLogger
+import qualified Mix.Plugin.Shell       as MixShell
 import qualified ScrapBook
 import qualified Version
 
@@ -28,17 +30,25 @@ main :: IO ()
 main = withGetOpt' "[options] [input-file]" opts $ \r args usage ->
   if | r ^. #help    -> hPutBuilder stdout (fromString usage)
      | r ^. #version -> hPutBuilder stdout (Version.build version)
-     | otherwise     -> runCmd r $ listToMaybe args
+     | otherwise     -> runCmd r (listToMaybe args)
   where
-    opts = #help    @= helpOpt
-        <: #version @= versionOpt
-        <: #verbose @= verboseOpt
+    opts = #help       @= helpOpt
+        <: #version    @= versionOpt
+        <: #verbose    @= verboseOpt
+        <: #skip       @= skipOpt
+        <: #withCopy   @= withCopyOpt
+        <: #withCommit @= withCommitOpt
+        <: #withPush   @= withPushOpt
         <: nil
 
 type Options = Record
-  '[ "help"    >: Bool
-   , "version" >: Bool
-   , "verbose" >: Bool
+  '[ "help"       >: Bool
+   , "version"    >: Bool
+   , "verbose"    >: Bool
+   , "skip"       >: Bool
+   , "withCopy"   >: Bool
+   , "withCommit" >: Bool
+   , "withPush"   >: Bool
    ]
 
 helpOpt :: OptDescr' Bool
@@ -50,9 +60,22 @@ versionOpt = optFlag [] ["version"] "Show version"
 verboseOpt :: OptDescr' Bool
 verboseOpt = optFlag ['v'] ["verbose"] "Enable verbose mode: verbosity level \"debug\""
 
+skipOpt :: OptDescr' Bool
+skipOpt = optFlag [] ["skip"] "Skip generate HTML"
+
+withCopyOpt :: OptDescr' Bool
+withCopyOpt = optFlag [] ["with-copy"] "Copy files by another branch before generate HTML"
+
+withCommitOpt :: OptDescr' Bool
+withCommitOpt = optFlag [] ["with-commit"] "Create commit after generate HTML"
+
+withPushOpt :: OptDescr' Bool
+withPushOpt = optFlag [] ["with-push"] "Push commit after create commit"
+
 type Env = Record
   '[ "logger" >: LogFunc
    , "config" >: Config
+   , "work"   >: FilePath
    ]
 
 runCmd :: Options -> Maybe FilePath -> IO ()
@@ -62,8 +85,14 @@ runCmd opts (Just path) = do
   let plugin = hsequence
              $ #logger <@=> MixLogger.buildPlugin logOpts
             <: #config <@=> pure config
+            <: #work   <@=> pure "."
             <: nil
-  Mix.run plugin $ generate path
+  Mix.run plugin $ do
+    when (opts ^. #withCommit) $ MixShell.exec (Git.pull [])
+    when (opts ^. #withCopy)   $ copyFilesByAnotherBranch
+    when (not $ opts ^. #skip) $ generate path
+    when (opts ^. #withCommit) $ commitGeneratedFiles
+    when (opts ^. #withPush)   $ pushCommit
   where
     logOpts = #handle @= stdout
            <: #verbose @= (opts ^. #verbose)
@@ -107,3 +136,27 @@ handler e = MixLogger.logError (displayShow e) >> pure []
 
 feed' :: ScrapBook.Format
 feed' = embedAssoc $ #feed @= ()
+
+copyFilesByAnotherBranch :: RIO Env ()
+copyFilesByAnotherBranch = do
+  targets <- view #copy <$> asks (gitConfig . view #config)
+  MixLogger.logDebug $ "copy files by " <> displayShow targets
+  MixShell.exec $ forM_ (splitCopyTarget <$> targets) $ \(branch, path) ->
+    Git.checkout [branch, "--", path]
+
+commitGeneratedFiles :: RIO Env ()
+commitGeneratedFiles = do
+  files <- view #files <$> asks (gitConfig . view #config)
+  MixLogger.logDebug $ "create commit with " <> displayShow files
+  MixShell.exec $ do
+    Git.add files
+    changes <- Git.diffFileNames ["--staged"]
+    when (not $ null changes) $ Git.commit ["-m", message]
+  where
+    message = "[skip ci] Update planet haskell. See https://haskell.jp/antenna/ for new entries!"
+
+pushCommit :: RIO Env ()
+pushCommit = do
+  branch <- view #branch <$> asks (gitConfig . view #config)
+  MixLogger.logDebug $ "push commit to origin/" <> display branch
+  MixShell.exec (Git.push ["origin", branch])